# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion PartList",
        "contact": {
            "email": "antony.rodriguez@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionPartList/public/index.php/buscar_codiProy": {
            "post": {
                "tags": [
                    "Gestion Proyecto Zona"
                ],
                "summary": "Buscar el codigo proyecto",
                "operationId": "App\\Http\\Controllers\\ProyectozonaController::buscar_codiProy",
                "parameters": [
                    {
                        "name": "varCodiProy",
                        "in": "path",
                        "description": "Ingresar el codigo proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "varCodiProy": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "varCodiProy": "18-0097PIR"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Muestra el id de varCodiProy"
                    },
                    "407": {
                        "description": "No se encontro el Codigo del proyecto."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionPartList/public/index.php/regi_proy_zona": {
            "post": {
                "tags": [
                    "Gestion Proyecto Zona"
                ],
                "summary": "Registrar el proyecto zona",
                "operationId": "App\\Http\\Controllers\\ProyectozonaController::regi_proy_zona",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "Ingrese el id de proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varDescripTarea",
                        "in": "path",
                        "description": "Ingrese la descripcion de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "Ingrese id tipo producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varDescrip",
                        "in": "path",
                        "description": "Ingrese la descripcion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acti_usua",
                        "in": "path",
                        "description": "Ingrese la codigo del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "varDescripTarea": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varDescrip": {
                                        "type": "string"
                                    },
                                    "varCodigoPaquete": {
                                        "type": "string"
                                    },
                                    "acti_usua": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "1205",
                                    "varDescripTarea": "101",
                                    "intIdTipoProducto": "1",
                                    "varDescrip": "I003",
                                    "varCodigoPaquete": "I003"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "si existe entonces mandamos el id de ese id proyecto zona"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionPartList/public/index.php/vali_proy_zona": {
            "post": {
                "tags": [
                    "Gestion Proyecto Zona"
                ],
                "summary": "validar si la zona existe, en caso que exista mandara el id  proyecto zona",
                "operationId": "App\\Http\\Controllers\\ProyectozonaController::vali_proy_zona",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "Ingresar el id del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "intIdTipoProducto",
                        "in": "path",
                        "description": "Ingresar el id del tipo producto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varDescrip",
                        "in": "path",
                        "description": "Ingresar la descripcion de la zona",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdTipoProducto": {
                                        "type": "string"
                                    },
                                    "varDescrip": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "126",
                                    "0": "intIdTipoProducto",
                                    "varDescrip": "19_0021_1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exito. Ya que el id proyecto zona no existe."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}